+++ jaguar_serializer/example/user/book/book_mongo.jser.dart
@@ -9,31 +9,32 @@ part of 'book_mongo.dart';
 abstract class _$BookMongoSerializer implements Serializer<Book> {
   final _mongoId = const MongoId();
 
+  @override
   Map<String, dynamic> toMap(Book model,
       {bool withType: false, String typeKey}) {
     Map<String, dynamic> ret;
     if (model != null) {
       ret = <String, dynamic>{};
-      setNullableValue(ret, "_id", _mongoId.serialize(model.id));
-      setNullableValue(ret, "N", model.name);
-      setNullableValue(ret, "publishedYear", model.publishedYear);
+      setNullableValue(ret, '_id', _mongoId.serialize(model.id));
+      setNullableValue(ret, 'N', model.name);
+      setNullableValue(ret, 'publishedYear', model.publishedYear);
       setTypeKeyValue(typeKey, modelString(), withType, ret);
     }
     return ret;
   }
 
-  Book fromMap(Map<String, dynamic> map, {Book model, String typeKey}) {
+  @override
+  Book fromMap(Map<String, dynamic> map, {Book model}) {
     if (map == null) {
       return null;
     }
-    if (model is! Book) {
-      model = new Book();
-    }
-    model.id = _mongoId.deserialize(map["_id"] as ObjectId);
-    model.name = map["N"] as String;
-    model.publishedYear = map["publishedYear"] as int;
-    return model;
+    final obj = model ?? new Book();
+    obj.id = _mongoId.deserialize(map['_id'] as ObjectId);
+    obj.name = map['N'] as String;
+    obj.publishedYear = map['publishedYear'] as int;
+    return obj;
   }
 
-  String modelString() => "Book";
+  @override
+  String modelString() => 'Book';
 }