+++ jaguar_serializer/lib/src/annotations/annotations.dart
@@ -1,5 +1,3 @@
-library jaguar_serializer.annotations;
-
 import '../serializer/serializer.dart';
 import '../serializer/field_processor.dart';
 
@@ -43,7 +41,7 @@ class GenSerializer {
   /// final model = serializer.fromMap(map);
   /// print(model.myField); // print 'foo'
   /// ```
-  final FieldFormat fieldFormat;
+  final String fieldFormat;
 
   const GenSerializer(
       {this.fields: const <String, Property>{},
@@ -52,7 +50,7 @@ class GenSerializer {
       this.modelName,
       this.includeByDefault: true,
       this.nullableFields: true,
-      this.fieldFormat: FieldFormat.none});
+      this.fieldFormat});
 }
 
 class Property<T> {
@@ -165,5 +163,21 @@ const Property nonNullable = const Property<dynamic>(isNullable: false);
 const Property useConstructorForDefaultsValue =
     const Property<dynamic>(valueFromConstructor: true, isNullable: false);
 
-/// Determine the outpu format for a field
-enum FieldFormat { none, camelCase, snakeCase, kebabCase }
+/// Determine the output format for a field
+/// Example:
+///     class Foo {
+///         String camelCase = "bar";
+///     }
+///
+///     @GenSerializer(fieldFormat: FieldFormat.snakeCase)
+///     class FooSerializer ...
+///
+/// will produce
+///    {
+///       "camel_case": "bar"
+///    }
+class FieldFormat {
+  static const camelCase = 'camelCase';
+  static const snakeCase = 'snakeCase';
+  static const kebabCase = 'kebabCase';
+}