+++ jaguar_serializer_cli/lib/src/helpers/parser_helper.dart
@@ -139,9 +139,9 @@ class AnnotationParser {
     _parseModelType();
     _parseIncludeByDefault();
     _parseModelString();
-    _parseFields();
     _parseIgnore();
     _parseSerializers();
+    _parseFields();
     _parseFieldFormatters();
 
     final ret = new SerializerInfo(name, modelType,
@@ -164,9 +164,7 @@ class AnnotationParser {
   }
 
   void _parseFieldFormatters() {
-    final index =
-        obj.peek('fieldFormat').objectValue.getField("index").toIntValue();
-    final format = FieldFormat.values[index];
+    final format = obj.peek('fieldFormat')?.stringValue;
     fieldFormatter = (str) => str;
     if (format == FieldFormat.camelCase) {
       fieldFormatter = toCamelCase;
@@ -207,10 +205,12 @@ class AnnotationParser {
   void _parseFields() {
     final Map<DartObject, DartObject> map = obj.peek('fields')?.mapValue ?? {};
     map.forEach((DartObject dKey, DartObject dV) {
+      final key = dKey.toStringValue();
       if (isIgnore.isAssignableFromType(dV.type)) {
-        to[dKey.toStringValue()] = null;
-        from[dKey.toStringValue()] = null;
-      } else {
+        to[key] = null;
+        from[key] = null;
+      } else if (!to.containsKey(key) && !from.containsKey(key)) {
+        // if contains key it means they are ignored in _parseIgnore
         _processField(dKey, dV);
       }
     });
@@ -329,7 +329,9 @@ class AnnotationParser {
       }
     });
 
-    el.fields.where((f) => f.isFinal && !f.isStatic).forEach((FieldElement f) {
+    el.fields
+        .where((f) => f.isFinal && !f.isStatic && !f.isPrivate)
+        .forEach((FieldElement f) {
       mod.addFrom(
           new Field(f.displayName, f.type as InterfaceType, isFinal: true));
     });