+++ jaguar_serializer_cli/test/common/models/address_book/address_book.jser.dart
@@ -9,65 +9,66 @@ part of 'address_book.dart';
 abstract class _$Person implements Serializer<Person> {
   final _address = new Address();
 
+  @override
   Map<String, dynamic> toMap(Person model,
       {bool withType: false, String typeKey}) {
     Map<String, dynamic> ret;
     if (model != null) {
       ret = <String, dynamic>{};
-      setNullableValue(ret, "name", model.name);
-      setNullableValue(ret, "address",
+      setNullableValue(ret, 'name', model.name);
+      setNullableValue(ret, 'address',
           _address.toMap(model.address, withType: withType, typeKey: typeKey));
       setTypeKeyValue(typeKey, modelString(), withType, ret);
     }
     return ret;
   }
 
-  Person fromMap(Map<String, dynamic> map, {Person model, String typeKey}) {
+  @override
+  Person fromMap(Map<String, dynamic> map, {Person model}) {
     if (map == null) {
       return null;
     }
-    if (model is! Person) {
-      model = new Person();
-    }
-    model.name = map["name"] as String;
-    model.address = _address.fromMap(map["address"] as Map<String, dynamic>,
-        typeKey: typeKey);
-    return model;
+    final obj = model ?? new Person();
+    obj.name = map['name'] as String;
+    obj.address = _address.fromMap(map['address'] as Map<String, dynamic>);
+    return obj;
   }
 
-  String modelString() => "Person";
+  @override
+  String modelString() => 'Person';
 }
 
 abstract class _$Address implements Serializer<Address> {
+  @override
   Map<String, dynamic> toMap(Address model,
       {bool withType: false, String typeKey}) {
     Map<String, dynamic> ret;
     if (model != null) {
       ret = <String, dynamic>{};
-      setNullableValue(ret, "street", model.street);
-      setNullableValue(ret, "houseNum", model.houseNum);
-      setNullableValue(ret, "city", model.city);
-      setNullableValue(ret, "country", model.country);
-      setNullableValue(ret, "pincode", model.pincode);
+      setNullableValue(ret, 'street', model.street);
+      setNullableValue(ret, 'houseNum', model.houseNum);
+      setNullableValue(ret, 'city', model.city);
+      setNullableValue(ret, 'country', model.country);
+      setNullableValue(ret, 'pincode', model.pincode);
       setTypeKeyValue(typeKey, modelString(), withType, ret);
     }
     return ret;
   }
 
-  Address fromMap(Map<String, dynamic> map, {Address model, String typeKey}) {
+  @override
+  Address fromMap(Map<String, dynamic> map, {Address model}) {
     if (map == null) {
       return null;
     }
-    if (model is! Address) {
-      model = new Address();
-    }
-    model.street = map["street"] as String;
-    model.houseNum = map["houseNum"] as String;
-    model.city = map["city"] as String;
-    model.country = map["country"] as String;
-    model.pincode = map["pincode"] as String;
-    return model;
+    final obj = model ?? new Address();
+    obj.street = map['street'] as String;
+    obj.houseNum = map['houseNum'] as String;
+    obj.city = map['city'] as String;
+    obj.country = map['country'] as String;
+    obj.pincode = map['pincode'] as String;
+    return obj;
   }
 
-  String modelString() => "Address";
+  @override
+  String modelString() => 'Address';
 }