+++ jaguar_serializer_cli/test/common/models/bibliotek/author.dart
@@ -1,5 +1,5 @@
 import 'package:jaguar_serializer/jaguar_serializer.dart';
-part 'author.g.dart';
+part 'author.jser.dart';
 
 @GenSerializer(serializers: const [])
 class AuthorSerializer extends Serializer<Author> with _$AuthorSerializer {}
@@ -9,6 +9,7 @@ class Author {
   /// Name of the author
   String name;
 
+  @override
   bool operator ==(Object other) {
     if (other is Author) {
       return name == other.name;
@@ -16,4 +17,7 @@ class Author {
 
     return false;
   }
+
+  @override
+  int get hashCode => name.hashCode;
 }