+++ jaguar_serializer_cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.0.0-dev8
+
+- fix flutter --release build
+- support generics
+
+## 1.0.0-dev7
+
+- fix issues on static and private field
+
 ## 1.0.0-dev6
 
 - update source_gen and build